Moves things to lib directory.
