+++ client.src/features/views/gist-view/index.js
@@ -7,8 +7,8 @@ import * as _ from 'underscore';
 import * as bb from 'backbone';
 import * as mn from 'marionette';
 import * as Radio from 'radio';
-import GistbookView from '../gistbook-view';
-import Gist from '../../../features/entities/gist';
+import GistbookView from '../../../lib/gistbook-view';
+import Gist from '../../entities/gist';
 import ExistingMenu from './existing-menu';
 import gistbookUtil from '../../../util/gistbook-util';
