+++ client.src/lib/auth/index.js
