+++ client.src/lib/gistbook-view/gistbook-view.jst
