+++ client.src/lib/gistbook-view/managers/code-manager.js
