+++ client.src/lib/gistbook-view/managers/compiler.js
