+++ client.src/lib/gistbook-view/views/output-view/index.js
