+++ client.src/lib/gistbook-view/views/output-view/output-view.jst
