+++ client.src/lib/gistbook-view/views/output/document-view/document-view.jst
