+++ client.src/lib/gistbook-view/views/output/document-view/index.js
