+++ client.src/lib/gistbook-view/views/text/display/index.js
