+++ client.src/lib/gistbook-view/views/text/edit/index.js
