+++ client.src/lib/gistbook-view/views/wrappers/display-wrapper/display-wrapper.jst
