+++ client.src/lib/modal/index.js
