+++ client.src/routes/about.js
@@ -4,9 +4,10 @@
 
 import * as mn from 'marionette';
 import * as Radio from 'radio';
+import Route from '../lib/routing/route';
 import AboutView from '../features/views/about-view';
 
-export default mn.Route.extend({
+export default Route.extend({
   show: function() {
     Radio.command('rootView', 'showIn:container', new AboutView());
   }
