+++ client.src/routes/home.js
@@ -5,10 +5,11 @@
 
 import * as mn from 'marionette';
 import * as Radio from 'radio';
+import Route from '../lib/routing/route';
 import Gist from '../features/entities/gist';
 import GistView from '../features/views/gist-view';
 
-export default mn.Route.extend({
+export default Route.extend({
   show: function() {
     var gistView = new GistView({
       model: new Gist(),
