+++ client.src/routes/logout.js
@@ -4,10 +4,11 @@
 
 import * as mn from 'marionette';
 import * as Radio from 'radio';
+import Route from '../lib/routing/route';
 
 var authChannel = Radio.channel('auth');
 
-export default mn.Route.extend({
+export default Route.extend({
   redirect: function() {
     authChannel.command('logout');
     return '';
