+++ client.src/routes/settings.js
@@ -4,11 +4,12 @@
 
 import * as mn from 'marionette';
 import * as Radio from 'radio';
+import Route from '../lib/routing/route';
 import SettingsView from '../features/views/settings-view';
 
 var user = Radio.request('user', 'user');
 
-export default mn.Route.extend({
+export default Route.extend({
 
   // Redirect us to the home page if we're unauthorized
   // This is the only page that requires that you be authorized
