+++ client.src/routes/terms.js
@@ -4,9 +4,10 @@
 
 import * as mn from 'marionette';
 import * as Radio from 'radio';
+import Route from '../lib/routing/route';
 import TermsView from '../features/views/terms-view';
 
-export default mn.Route.extend({
+export default Route.extend({
   show: function() {
     Radio.command('rootView', 'showIn:container', new TermsView());
   }
